ugc/requirements.txt
fastapi==0.92.0
pydantic==1.10.5
orjson==3.8.7

motor==3.1.2

uvicorn[standard]==0.21.1
gunicorn==20.1.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"

dependency-injector==4.41.0
python-jose[cryptography]==3.3.0
sentry-sdk==1.25.0